Abstracting core settings out
